requirements.txt
beautifulsoup4==4.11.1
cloudscraper==1.2.60
pycaw==20220416
python-dotenv==0.20.0
comtypes==1.1.11
